added time header component
