+++ foqos/Views/HomeView.swift
@@ -80,28 +80,10 @@ struct HomeView: View {
             VStack(alignment: .leading, spacing: 20) {
                 if !profiles.isEmpty {
                     VStack(alignment: .leading, spacing: 10) {
-                        Text(timeString(from: strategyManager.elapsedTime))
-                            .font(.system(size: 80))
-                            .fontWeight(.semibold)
-                            .foregroundColor(
-                                isBlocking ? Color(hex: "#32CD32") : .primary
-                            )
-                            .opacity(isBlocking ? opacityValue : 1)
-                            .animation(
-                                .easeInOut(duration: 0.7).repeatForever(),
-                                value: opacityValue
-                            )
-                            .onChange(of: isBlocking) { _, newValue in
-                                if newValue {
-                                    withAnimation(
-                                        .easeInOut(duration: 1).repeatForever()
-                                    ) {
-                                        opacityValue = 0.3
-                                    }
-                                } else {
-                                    opacityValue = 1
-                                }
-                            }
+                        TimeHeader(
+                            elapsedTime: strategyManager.elapsedTime,
+                            isBlocking: isBlocking
+                        )
                     }
                     
                     VStack(alignment: .leading, spacing: 10) {
@@ -301,13 +283,6 @@ struct HomeView: View {
     
     private func unloadApp() {
         strategyManager.stopTimer()
-    }
-    
-    private func timeString(from timeInterval: TimeInterval) -> String {
-        let hours = Int(timeInterval) / 3600
-        let minutes = Int(timeInterval) / 60 % 60
-        let seconds = Int(timeInterval) % 60
-        return String(format: "%02d:%02d:%02d", hours, minutes, seconds)
     }
     
     private func showErrorAlert(message: String) {
